test: print logical instruction count per program
[ upstream commit db0e2c287bb2f50d951d8b118aae79f6755d4032 ]

As it says on the tin, echo instruction count for all programs, both if the
Collection fails to load, as well as at the end of the last line of the
verifier log on a successful load.

Example output of successful load:
```
=== RUN   TestVerifier/bpf_host_1
    verifier_test.go:216: tail_srv6_reply: processed 277 insns (limit 1000000) ... mark_read 8 (201 unverified insns)
...
```

Example output of unsuccessful load:

```
=== RUN   TestVerifier/bpf_lxc_1
    verifier_test.go:201: Full verifier log at bpf_lxc_1_verifier.log
    verifier_test.go:203: BPF unverified instruction count per program:
    verifier_test.go:205: tail_handle_ipv4: 1048 insns
```

Also output the original `err` containing the name of the prog that caused
the verifier error, which was accidentally removed by 18339641b5
("test/verifier: don't dump full log into test output"). It's now
documented why it's included.

Signed-off-by: Timo Beckers <timo@isovalent.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>