gha: fix waiting for images in conformance-gingko
[ upstream commit e1f838977dda2ed6e07dc18e1d30e05df21de4ed ]

Let's add an explicit dependency to the wait-for-images job, so that
subsequent jobs are not started until the images are actually available.
This also prevents starting the full matrix jobs if the images don't
get ready in time. Finally, let's adapt status reporting, so that
we properly report a failure in that case.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>