cilium: debugging InNoState errors on agent roll
When rolling cilium-agent while under stress test with encryption a
small number of NoStateIn are seen. When we inspect the pod both
state rules for the oldIP and newIP are installed, but the error
counter is incrementing really close to pod init. Close enough
that I can't get a dump of the xfrm state before/during when the
errors occur. Best I can say is the errors are happening within
the first <10s of the pod being up.

Further complicating reproduction is I need to have a stress
test and then restart the cilium-agent pod a few times to even
catch a small number of errors.

The working theory is somehow remote nodes learn the new hostIP
and start sending encrypted traffic with that outer IP before
the datapath is plumbed on the node.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>